+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www"
+    "start": "nodemon ./BackEnd/index.js"
   },
   "dependencies": {
     "body-parser": "^1.18.3",
